+++ src/ReactorUI.Skia/Framework/Button.cs
@@ -13,7 +13,7 @@ namespace ReactorUI.Skia.Framework
 
         #region Public Events
         public event EventHandler<Input.MouseEventArgs> Click;
-        protected virtual void OnClick(int x, int y, Input.MouseEventsContext context)
+        protected virtual void OnClick(double x, double y, Input.MouseEventsContext context)
         {
             Click?.Invoke(this, new Input.MouseEventArgs(context.MouseButtons, context.Clicks, x, y, context.Delta));
         }
@@ -21,7 +21,7 @@ namespace ReactorUI.Skia.Framework
 
         bool _mouseDown = false;
 
-        protected override void OnMouseDown(int x, int y, MouseEventsContext context)
+        protected override void OnMouseDown(double x, double y, MouseEventsContext context)
         {
             _mouseDown = true;
             context.CaptureTo = this;
@@ -29,7 +29,7 @@ namespace ReactorUI.Skia.Framework
             base.OnMouseDown(x, y, context);
         }
 
-        protected override void OnMouseUp(int x, int y, MouseEventsContext context)
+        protected override void OnMouseUp(double x, double y, MouseEventsContext context)
         {
             if (_mouseDown)
             {
