+++ src/ReactorUI.Skia/Framework/Input/MouseEventArgs.cs
@@ -7,7 +7,7 @@ namespace ReactorUI.Skia.Framework.Input
     public class MouseEventArgs : EventArgs
     {
         [System.Diagnostics.DebuggerStepThrough]
-        public MouseEventArgs(MouseButtons mouseButtons, int clicks, int x, int y, int delta)
+        public MouseEventArgs(MouseButtons mouseButtons, int clicks, double x, double y, int delta)
         {
             MouseButtons = mouseButtons;
             Clicks = clicks;
@@ -18,8 +18,8 @@ namespace ReactorUI.Skia.Framework.Input
 
         public MouseButtons MouseButtons { get; }
         public int Clicks { get; }
-        public int X { get; }
-        public int Y { get; }
+        public double X { get; }
+        public double Y { get; }
         public int Delta { get; }
     }
 }
