+++ src/ReactorUI/Primitives/Size.cs
@@ -45,7 +45,7 @@ namespace ReactorUI.Primitives
         public bool IsCloseTo(Size other) =>
             Math.Abs(Width - other.Width) < 1e-10 && Math.Abs(Height - other.Height) < 1e-10;
 
-        public bool Contains(int x, int y) => !IsEmpty && x >= 0 && y >= 0 && x <= Width && y <= Height;
+        public bool Contains(double x, double y) => !IsEmpty && x >= 0 && y >= 0 && x <= Width && y <= Height;
 
     }
 }
